requirements.txt
crewai
crewai[tools]
litellm
yfinance
streamlit
pandas
ta
lxml
python-dotenv
duckduckgo-search
requests